Factor out ResourceVoteControls component
